Bump DMA_BUFSIZE to avoid overruns in DMA_BlockRead
Lands of Lore: Guardians of Destiny running in DOSBox on Chrome on a
2.9Ghz Core i9 frequently runs into this problem.

Occasionally when SBLASTER_CallBack runs, the resulting call into
DMA_BlockRead from GenerateDMASound will read more than 2048 bytes
(1024 * sizeof(Bit16s)) into sb.dma.buf.b16. The resulting overrun
corrupts memory and leads to segfaults.

This doesn't happen when DOSBox is running as a native binary outside
of the browser on the same machine. It isn't clear why this would be
the case, but perhaps the tick rate is higher, and results DmaChannel
reading less data.

There shouldn't be any side effects from bumping the buffer size. The
buffer is written out to mixer.work, which appears to be large enough
already.